Day 10/backend/requirements.txt

# # LangChain ecosystem
# langchain==0.1.20
# langchain-core==0.1.53
# langchain-community==0.0.38
# langchain-google-genai==0.0.9

# # # Google Generative AI SDK
# # google-generativeai==0.5.4

# # Embeddings + Vector store
# chromadb==0.4.24
# sentence-transformers==2.6.1
# huggingface-hub==0.23.0

# # Document parsing
# pypdf==4.2.0
# pdf2image==1.17.0
# pytesseract==0.3.10
# python-pptx==0.6.23

langchain
langchain-community
langchain-google-genai
tavily-python
sentence-transformers
chromadb
python-dotenv
pdf2image
pytesseract
python-pptx

